% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenM-data.R
\docType{data}
\name{eigenM}
\alias{eigenM}
\title{An example of grid points and eigendecomposition for reprodcing kernel of cubic spline.}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(eigenM)
}
\description{
It contains \code{e} a list of eigenvalues and eigenfunctions at grid points, \code{xg} a vector of 1000 evenly spaced points.
}
\examples{
data(eigenM)
\dontrun{
# check the number of eigenvalues
length(eigenM$e$values)
}
}
\keyword{datasets}
