% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrspline.cubic.R
\name{lrspline.cubic}
\alias{lrspline.cubic}
\title{Low-rank Approximation Based on Eigenspaces for Cubic Smoothing Spline Estimates}
\usage{
lrspline.cubic(x, y, xg, e, K = 30, method = "REML", pstd = FALSE)
}
\arguments{
\item{x}{The values of independent variable. It should be a vector.}

\item{y}{The values of dependent variable. It should be a vector.}

\item{xg}{The gridpoints used for approximation of}

\item{e}{A list of two elements of "values" and "vectors", which refer respectively, the eigenvalues and eigenfunctions of reproducing kernel for cubic splines at the pre-selected gridpoints (must agree with xg).}

\item{K}{An integer value. The truncation parameter indicates the number of eigenvalues/eigenfunctions used in approximation. The default value is 30.}

\item{method}{A character string. If "REML" the LME model is fit by maximizing the restricted log-likelihood. If "ML" the log-likelihood is maximized. Defaults to "REML".}

\item{pstd}{An indicator of whether standard deviation is desired. The default value is FALSE.}
}
\value{
A vector(s) of following component(s):
\item{fit}{The low-rank approximation of cubic smoothing spline estimate.}
\item{pstd}{The corrsponding posterior standard deviation.}
}
\description{
Computes a low-rank approximation based on eigenspaces, where estimation utilizes functions for linear mixed effect model (LME).
}
\examples{
\dontrun{
data(eigenM)
x <- runif(1000)
y <- sin(32*pi*x)-8*(x-.5)^2 + rnorm(1000)
lrspline.cubic(x,y,xg,e,K,method="REML",pstd=FALSE)
}
}
