% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nystrom.cubic.R
\name{nystrom.cubic}
\alias{nystrom.cubic}
\title{Nystrom Approximation for Cubic Smoothing Spline Estimates}
\usage{
nystrom.cubic(x, y, p = 30, method = "REML", pstd = FALSE)
}
\arguments{
\item{x}{The values of independent variable. It should be a vector.}

\item{y}{The values of dependent variable. It should be a vector.}

\item{p}{An integer value. The selection parameter indicates the number of columns for random selection and approximation. The default value is 30.}

\item{method}{A character string. If "REML" the LME model is fit by maximizing the restricted log-likelihood. If "ML" the log-likelihood is maximized. Defaults to "REML".}

\item{pstd}{An indicator of whether standard deviation is desired. The default value is FALSE.}

\item{xg}{The gridpoints used for approximation of}

\item{e}{A list of two elements of "values" and "vectors", which refer respectively, the eigenvalues and eigenfunctions of reproducing kernel for cubic splines at the pre-selected gridpoints (must agree with xg).}
}
\value{
A vector(s) of following component(s):
\item{fit}{The Nystrom approximation of cubic smoothing spline estimate.}
\item{pstd}{The corrsponding posterior standard deviation.}
}
\description{
Computes a Nystrom approximation based on randomly selected columns of $\Sigma$ matrix, where estimation utilizes functions for linear mixed effect model (LME).
}
\examples{
\dontrun{
data(eigenM)
x <- runif(1000)
y <- sin(32*pi*x)-8*(x-.5)^2 + rnorm(1000)
nystrom.cubic(x,y,xg,e,K,method="REML",pstd=FALSE)
}
}
